# Improve unpackString performance