@miekg 2019-09-21
Escapes are relatively rare as well, so might be the wrong target to
optimize for

On Sat, 21 Sep 2019, 09:33 Codecov, <notifications@github.com> wrote:

> Codecov <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=h1> Report
>
> Merging #1011 <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=desc>
> into master
> <https://codecov.io/gh/miekg/dns/commit/ba5bfd029554f8f9c5215e84160d49fd02b8a397?src=pr&el=desc>
> will *increase* coverage by 0.04%.
> The diff coverage is 92.85%.
>
> [image: Impacted file tree graph]
> <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=tree>
>
> @@            Coverage Diff            @@
> ##           master   #1011      +/-   ##
>
> =========================================
> + Coverage   54.66%   54.7%   +0.04%
>
> =========================================
>
>   Files          41      41
>
>   Lines        9844    9854      +10
>
> =========================================
> + Hits         5381    5391      +10
>
>   Misses       3437    3437
>
>   Partials     1026    1026
>
> Impacted Files <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=tree> Coverage
> Δ
> msg_helpers.go
> <https://codecov.io/gh/miekg/dns/pull/1011/diff?src=pr&el=tree#diff-bXNnX2hlbHBlcnMuZ28=> 58.46%
> <92.85%> (+0.95%) ⬆️
> ------------------------------
>
> Continue to review full report at Codecov
> <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=continue>.
>
> *Legend* - Click here to learn more
> <https://docs.codecov.io/docs/codecov-delta>
> Δ = absolute <relative> (impact), ø = not affected, ? = missing data
> Powered by Codecov
> <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=footer>. Last update
> ba5bfd0...bb8072a
> <https://codecov.io/gh/miekg/dns/pull/1011?src=pr&el=lastupdated>. Read
> the comment docs <https://docs.codecov.io/docs/pull-request-comments>.
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/miekg/dns/pull/1011?email_source=notifications&email_token=AACWIW6ETHPDFCTGSNCRK7LQKXL5NA5CNFSM4IY5OLLKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD7INL3Q#issuecomment-533779950>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AACWIWZG4OLQOCYKCBVITGDQKXL5NANCNFSM4IY5OLLA>
> .
>