@miekg 2019-09-21
~6- 10 % faster is nothing to sneeze at, /lgtm